Move from lazy_static to a weak static in C lib
Sadly, a C library is necessary for this, as `#[linkage="weak"]` is still
unstable in Rust.

Sadly, much like with the lazy-static mutex implementation, it is a
wide-hitting hammer and will also hit platforms which aren’t affected by
this (by virtue of using thread-local dlerrors).

This will require a "major" version bump.

Fixes #32