Do not encourage use of libloading for TLS
On MSVC Windows loading a TLS symbol will give an address to the
initializer (which when overwritten simply changes the default value new
threads gain for the variable).

On macOS thread locals are misimplemented entirely.

On Linux things "just work", but unknown to what extent, really.

Since the behaviour of using thread local variables as if they were
plain global variables is mostly undefined, lets just not make any
claims of this working.